+++ songqueue.go
@@ -57,7 +57,7 @@ func (q *SongQueue) NextItem() {
 func (q *SongQueue) PeekNext() (string, string, error) {
 	var title, submitter string
 	if q.Len() > 1 {
-		if q.queue[1].ItemType == "playlist" {
+		if q.queue[1].ItemType() == "playlist" {
 			title = q.queue[1].(*Playlist).songs.queue[0].(*Song).title
 			submitter = q.queue[1].(*Playlist).submitter
 		} else {
@@ -65,7 +65,9 @@ func (q *SongQueue) PeekNext() (string, string, error) {
 			submitter = q.queue[1].(*Song).submitter
 		}
 		return title, submitter, nil
-	} else if q.CurrentItem().ItemType == "playlist" {
+	} else if q.Len() == 0 {
+		return "", "", errors.New("There is no item next in the queue.")
+	} else if q.CurrentItem().ItemType() == "playlist" {
 		title = q.CurrentItem().(*Playlist).songs.queue[1].(*Song).title
 		submitter = q.CurrentItem().(*Playlist).submitter
 		return title, submitter, nil